detect

+++ src/detection.py
@@ -21,9 +21,10 @@ def runDetection(image, extention):
     with open("./images/imgToGuessNew." + extention, "rb") as img_file:
       byte_content = img_file.read()
   
-  #remove file
+  #remove files
   os.remove("./images/imgToGuessNew." + extention)
-  os.remove("./images/imgToGuess." + extention)
+  if os.path.isfile("./images/imgToGuess." + extention):
+    os.remove("./images/imgToGuess." + extention)
   
   base64_bytes = b64encode(byte_content)
   base64_string = base64_bytes.decode(ENCODING)
